Update README.md
Suggest adding the line behind the python 3.8 that it will need to be installed by the user to run the project. I installed python 3.8 in the same directory as the cloned files on my pc and then ran using a powershell terminal. That might be a little more detail to add as well; to make sure python can run in the directory in which the user is running the files. Otherwise, everything looks good!